Level progression, different bricks.
Added level progression which counts the number of bricks on the screen and when all the bricks are gone, it moves to the next level. Also added different types of bricks. Red bricks need to be hit twice to be broken and blue bricks need to be hit three times to be broken. With this, single player is finally complete