Powerups, Pause menu
Added a pause menu. Hit escape while playing and it pauses the game. Has two buttons, one for continuing the game and the other to quit the game.

Also added two powerups. One powerup gives you a life back when the slider hits the heart. The other is a clone ball that helps get rid of bricks but the player doesn't lose a life if they miss the clone ball.