Shop
Changed the shop to make it completely functional.

Made several scripts.

Ball Skins.gd: Script to handle all the ball skins. In this script you can buy a skin and have that skin save even after the game is closed. Once a skin is bought, you can then select the skin to equip it onto the ball. When a skin is selected, it shows up in game with that skin.

Shop.gd: Handles all the events in the shop. Meaning buttons that can be pressed such as the purchase buttons and the return to main menu buttons. This also loads and saves the store so it remembers the state of the store even after closing the game. It will also check if the player has enough points to buy said item

Label.gd: This is just a script to show the price of the skins.